refactor: all buttons made out of the wooden sign are unified under WoodenButton
