--- src/common/BackButton/BackButton.css
@@ -1,17 +1,0 @@
-.backButtonDesign{
-    color: white;
-    cursor: pointer;
-    height: 2.3em;
-    width: 6.1em;
-    transition: 0.7s;
-    background-image: url('../../assets/botones.png');
-    background-position: center;
-    background-size: cover;
-    background-repeat: no-repeat;
-    filter: drop-shadow(0 0 0.75rem #ffa500);
-}
-.backText{
-    color: white;
-    filter: drop-shadow(0 0 0.2rem #f5e905, 0 0 0.5rem #ffa500);
-    height: 1em;
-}
