--- src/common/DeleteButton/DeleteButton.css
@@ -1,17 +1,0 @@
-.deleteButtonDesign{
-    color: white;
-    cursor: pointer;
-    height: 2.3em;
-    width: 6.1em;
-    transition: 0.7s;
-    background-image: url('../../assets/botonesCancelar.png');
-    background-position: center;
-    background-size: cover;
-    background-repeat: no-repeat;
-    filter: drop-shadow(0 0 0.4em #BF2E2E);
-}
-.deleteText{
-    color: white;
-    filter: drop-shadow(0 0 0.5rem #e35151, 0 0 0 0.2rem #BF2E2E);
-    height: 1em;
-}
