--- src/common/ModifyButton/ModifyButton.css
@@ -1,17 +1,0 @@
-.modifyButtonDesign{
-    color: white;
-    cursor: pointer;
-    height: 2.3em;
-    width: 6.1em;
-    transition: 0.7s;
-    background-image: url('../../assets/botonesEditar.png');
-    background-position: center;
-    background-size: cover;
-    background-repeat: no-repeat;
-    filter: drop-shadow(0 0 0.75rem #20acde);
-}
-.deleteText{
-    color: white;
-    filter: drop-shadow(0 0 0.2rem #274550, 0 0 0.5rem #7ebed5);
-    height: 1em;
-}
