+++ src/common/WoodenButton/WoodenButton.css
@@ -1,0 +1,53 @@
+.backButtonDesign{
+    color: white;
+    cursor: pointer;
+    height: 2.3em;
+    width: 6.1em;
+    transition: 0.7s;
+    background-image: url('../../assets/botones.png');
+    background-position: center;
+    background-size: cover;
+    background-repeat: no-repeat;
+    filter: drop-shadow(0 0 0.75rem #ffa500);
+}
+.backText{
+    color: white;
+    filter: drop-shadow(0 0 0.2rem #f5e905, 0 0 0.5rem #ffa500);
+    height: 1em;
+}
+
+.modifyButtonDesign{
+    color: white;
+    cursor: pointer;
+    height: 2.3em;
+    width: 6.1em;
+    transition: 0.7s;
+    background-image: url('../../assets/botonesEditar.png');
+    background-position: center;
+    background-size: cover;
+    background-repeat: no-repeat;
+    filter: drop-shadow(0 0 0.75rem #20acde);
+}
+.modifyText{
+    color: white;
+    filter: drop-shadow(0 0 0.2rem #274550, 0 0 0.5rem #7ebed5);
+    height: 1em;
+}
+
+.deleteButtonDesign{
+    color: white;
+    cursor: pointer;
+    height: 2.3em;
+    width: 6.1em;
+    transition: 0.7s;
+    background-image: url('../../assets/botonesCancelar.png');
+    background-position: center;
+    background-size: cover;
+    background-repeat: no-repeat;
+    filter: drop-shadow(0 0 0.4em #BF2E2E);
+}
+.deleteText{
+    color: white;
+    filter: drop-shadow(0 0 0.5rem #e35151, 0 0 0 0.2rem #BF2E2E);
+    height: 1em;
+}
